requirements.txt
customtkinter
pyautogui
requests
python-dotenv
opencv-python